requirements.txt
alpha-vantage==2.3.1
bernhard==0.2.6
black==22.12.0
bottle==0.12.23
cassandra-driver==3.25.0
chevron==0.14.0
couchdb==1.2
dask==2022.12.1
dm-tree==0.1.7
elasticsearch==8.5.3
enum34==1.1.10
financedatabase==1.0.2
finrl-meta==0.3.6
finta==1.3
flake8==6.0.0
fmpsdk==20221012.0
fundamentalanalysis==0.2.14
glances==3.3.0.4
graphitesender==0.11.2
influxdb==5.3.1
influxdb-client==1.35.0
kafka-python==2.0.2
keras-preprocessing==1.1.2
line-profiler==4.0.2
mathjax==0.1.2
memory-profiler==0.61.0
mlflow==2.1.1
nasdaq-data-link==1.0.4
netifaces==0.11.0
paho-mqtt==1.6.1
pandas-datareader==0.10.0
pika==1.3.1
pip-autoremove==0.10.0
pip-search==0.0.12
pip-upgrader==1.4.15
potsdb==1.0.3
py3nvml==0.2.7
pyenchant==3.2.2
pygal==3.0.0
pylint==2.15.8
pymdstat==0.4.2
pympler==1.0.1
pynvim==0.4.3
pyportfolioopt==1.5.4
pysnmp==4.4.12
pytest-depends==1.0.1
pytest-xdist==3.1.0
rope==1.6.0
scandir==1.10.0
seaborn-image==0.5.1
statsd==4.0.1
svgwrite==1.4.3
tensorflow==2.11.0
torchvision==0.14.1
tvdatafeed==2.1.0
wifi==0.3.8
xarray==2023.2.0
zeroconf==0.39.4
